Added more of the code needed for job submit APIs.

+++ src/api/cancel.c
@@ -51,7 +51,7 @@ int slurm_cancel_job ( uint32_t job_id )
 	{
 		case RESPONSE_SLURM_RC:
 			rc_msg = ( return_code_msg_t * ) response_msg . data ;
-			return rc_msg -> return_code ;
+			return (int) rc_msg -> return_code ;
 			break ;
 		default:
 			return SLURM_UNEXPECTED_MSG_ERROR ;
